Test ProductService delete

+++ backend/dscatalog/src/main/java/com/fabio/dscatalog/services/ProductService.java
@@ -59,6 +59,9 @@ public class ProductService {
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
         try{
             productRepository.deleteById(id);
         }catch(DataIntegrityViolationException exception){
